requirements.txt
asgiref==3.8.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-yasg==1.21.9
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2024.2
redis==5.2.1
sqlparse==0.5.3
sib-api-v3-sdk==7.6.0